arm: dts: add qcom-multi-dummy DT
This is a dummy DT containing some custom match data. It is expected to
be used as the default DTB along with a multi-DTB FIT image as the ramdisk
in a boot image, and allows for multiple different Qualcomm boards to be
supported by a single boot image.

To build, set CONFIG_DEFAULT_DEVICE_TREE="qcom-multi-dummy", and list
the DTBs to include with CONFIG_OF_LIVE.

follow the normal build instructions, using qcom-multi-dummy.dtb as the
board DT. When running mkbootimg, include

  --ramdisk out/fit-dtb.blob

Signed-off-by: Caleb Connolly <caleb.connolly@linaro.org>