boot: support booting TF-A from a FIT
U-Boot already supporting booting TF-A/OP-TEE/U-Boot all from a FIT
image, but only in SPL mode. Now that the code to do this has been moved
out of SPL, we can re-use it to enable booting this way from full fat
U-Boot.

This may be useful for development (particularly on boards that lack
SPL support) as it makes iterating on TF-A changes easy by allowing the
FIT image to be loaded via fastboot or some other mechanism.

Signed-off-by: Caleb Connolly <caleb.connolly@linaro.org>